Add map_as<T>.

+++ tests/src/blopp_tests_filesystem.cpp
@@ -2,7 +2,7 @@
 
 namespace {
     struct my_test_object {
-        size_t id = 0;
+        uint64_t id = 0;
         std::string name = {};      
     };
 }
